Add form-serialization classes to the 'query' module
