+++ lib/querying.py
@@ -25,234 +25,10 @@ from kontext import Kontext
 import corplib
 import plugins
 import l10n
-from query import CQLDetectWithin
+from query import (CQLDetectWithin, FilterFormArgs, QueryFormArgs, SortFormArgs, SampleFormArgs, ShuffleFormArgs,
+                   build_conc_form_args)
 from translation import ugettext as _
 from controller import exposed
-
-
-def has_tag_support(corpname):
-    return plugins.has_plugin('taghelper') and plugins.get('taghelper').tag_variants_file_exists(corpname)
-
-
-class ConcFormArgs(object):
-    """
-    A helper class to handle miscellaneous
-    form (filter, query, sort,...) args
-    properly. It is used only indirectly
-    - we create an instance and immediately
-    after that we export it (either to serialize
-    it via conc_persistence or to pass it
-    to the client-side).
-    """
-
-    def __init__(self, persist):
-        self._persistent = persist
-        self._op_key = '__new__'
-
-    def updated(self, attrs, op_key):
-        """
-        Return an updated self object
-        (the same instance). There must
-        be always the 'op_key' value
-        present to emphasize the fact
-        that only serialized data (i.e.
-        data with their database key)
-        can be used to update an 'unbound'
-        instance.
-        """
-        for k, v in attrs.items():
-            if k in vars(self):
-                setattr(self, k, v)
-        self._op_key = op_key
-        return self
-
-    def to_dict(self):
-        tmp = dict((k, v) for k, v in self.__dict__.items() if not k.startswith('_'))
-        if not self.is_persistent:
-            tmp['op_key'] = self._op_key
-        return tmp
-
-    @property
-    def is_persistent(self):
-        return self._persistent
-
-    @property
-    def op_key(self):
-        """
-        op_key property has a special status as
-        it is kept separate from other attributes
-        and is exported only if an instance is
-        persistent (= loaded from database). I.e.
-        newly created objects (where op_key == '__new__')
-        should not export it.
-        """
-        return self._op_key
-
-    def make_saveable(self):
-        self._op_key = '__new__'
-        self._persistent = True
-
-
-class LgroupOpArgs(ConcFormArgs):
-    """
-    This is used to store special actions that modify
-    compiled query string but are produced in a special
-    way and thus cannot be edited again (e.g. lines groups
-    operations).
-    """
-    def __init__(self, persist):
-        super(LgroupOpArgs, self).__init__(persist)
-        self.form_type = 'lgroup'
-
-
-class LockedOpFormsArgs(ConcFormArgs):
-    """
-    This is used to store actions that modify compiled
-    query string and are mapped to an existing user-editable
-    form (see the difference with LgroupOpArgs) but we
-    do not want user to edit them manually (e.g. user
-    filters manually selected lines which produces a bunch
-    of token IDs - nothing human-friendly). We actually
-    do not bother with storing the arguments.
-    """
-    def __init__(self, persist):
-        super(LockedOpFormsArgs, self).__init__(persist)
-        self.form_type = 'locked'
-
-
-class QueryFormArgs(ConcFormArgs):
-    """
-    QueryFormArgs collects arguments required
-    to initialize the 'first_form' for one or more
-    corpora.
-
-    The class is only used to make collecting and
-    serializing data easier. Stored data are expected
-    to be JSON-serializable.
-    """
-    def __init__(self, corpora, persist):
-        super(QueryFormArgs, self).__init__(persist)
-        self.form_type = 'query'
-        self.curr_query_types = dict((c, None) for c in corpora)
-        self.curr_queries = dict((c, None) for c in corpora)
-        self.curr_pcq_pos_neg_values = dict((c, None) for c in corpora)
-        self.curr_lpos_values = dict((c, None) for c in corpora)
-        self.curr_qmcase_values = dict((c, None) for c in corpora)
-        self.curr_default_attr_values = dict((c, None) for c in corpora)
-        self.tag_builder_support = dict((c, None) for c in corpora)
-        for corp in self.tag_builder_support.keys():
-            self.tag_builder_support[corp] = has_tag_support(corp)
-
-
-class FilterFormArgs(ConcFormArgs):
-    """
-    FilterFormArgs collects arguments required
-    to initialize the 'filter' form.
-
-    The class is only used to make collecting and
-    serializing data easier. Stored data are expected
-    to be JSON-serializable.
-    """
-    def __init__(self, maincorp, persist):
-        super(FilterFormArgs, self).__init__(persist)
-        self.form_type = 'filter'
-        self.query_type = 'iquery'
-        self.query = ''
-        self.maincorp = maincorp
-        self.pnfilter = 'p'
-        self.filfl = 'f'
-        self.filfpos = '-5'
-        self.filtpos = '5'
-        self.inclkwic = True
-        self.qmcase = False
-        self.default_attr = 'word'
-        self.tag_builder_support = has_tag_support(self.maincorp)
-
-
-class SortFormArgs(ConcFormArgs):
-    """
-    SortFormArgs collects arguments required
-    to initialize the 'sort' form.
-
-    The class is only used to make collecting and
-    serializing data easier. Stored data are expected
-    to be JSON-serializable.
-    """
-    def __init__(self, persist):
-        """
-        args:
-            persist -- specify whether the object should be stored
-                       to disk when the current action is finished
-        """
-        super(SortFormArgs, self).__init__(persist)
-        self.form_type = 'sort'
-        self.form_action = 'sortx'
-        self.sattr = ''
-        self.skey = 'kw'
-        self.spos = 3  # number of tokens to sort
-        self.sicase = ''
-        self.sbward = ''
-        self.sortlevel = 1
-        self.ml1attr = ''
-        self.ml2attr = ''
-        self.ml3attr = ''
-        self.ml4attr = ''
-        self.ml1icase = ''
-        self.ml2icase = ''
-        self.ml3icase = ''
-        self.ml4icase = ''
-        self.ml1bward = ''
-        self.ml2bward = ''
-        self.ml3bward = ''
-        self.ml4bward = ''
-        self.ml1pos = 1
-        self.ml2pos = 1
-        self.ml3pos = 1
-        self.ml4pos = 1
-        self.ml1ctx = u'0~0>0'
-        self.ml2ctx = u'0~0>0'
-        self.ml3ctx = u'0~0>0'
-        self.ml4ctx = u'0~0>0'
-
-
-class SampleFormArgs(ConcFormArgs):
-
-    def __init__(self, persist):
-        super(SampleFormArgs, self).__init__(persist)
-        self.form_type = 'sample'
-        self.rlines = '250'
-
-
-class ShuffleFormArgs(ConcFormArgs):
-
-    def __init__(self, persist):
-        super(ShuffleFormArgs, self).__init__(persist)
-        self.form_type = 'shuffle'
-
-
-def build_conc_form_args(data, op_key):
-    """
-    A factory method to create a conc form args
-    instance based on deserialized data from
-    conc_persistnece database.
-    """
-    tp = data['form_type']
-    if tp == 'query':
-        return QueryFormArgs(corpora=data.get('corpora', []), persist=False).updated(data, op_key)
-    elif tp == 'filter':
-        return FilterFormArgs(maincorp=data['maincorp'], persist=False).updated(data, op_key)
-    elif tp == 'sort':
-        return SortFormArgs(persist=False).updated(data, op_key)
-    elif tp == 'sample':
-        return SampleFormArgs(persist=False).updated(data, op_key)
-    elif tp == 'shuffle':
-        return ShuffleFormArgs(persist=False).updated(data, op_key)
-    elif tp == 'lgroup':
-        return LgroupOpArgs(persist=False).updated(data, op_key)
-    elif tp == 'locked':
-        return LockedOpFormsArgs(persist=False).updated(data, op_key)
-    else:
-        raise ValueError('Cannot determine stored conc args class from type %s' % (tp,))
 
 
 class Querying(Kontext):
